Move ConsensusCommitOutput into its own module
